Раскидала классы, enum'ы и интерфейсы по отдельным файлам
