+++ rpgsaga/saga/src/enums/carnivoreNames.ts
@@ -1,0 +1,5 @@
+// Список хищников, которые могут охотиться на Herbivore
+export enum CarnivoreNames {
+  wolf = 'wolf',
+  bear = 'bear',
+}
