+++ rpgsaga/saga/src/enums/plantNames.ts
@@ -1,0 +1,9 @@
+// Список растений, овощей и фруктов, которые ест PlantEater
+export enum PlantNames {
+  carrot = 'carrot',
+  grass = 'grass',
+  apple = 'apple',
+  pear = 'pear',
+  cucumber = 'cucumber',
+  cabbage = 'cabbage',
+}
