+++ rpgsaga/saga/src/interfaces/herbivore.ts
@@ -1,0 +1,7 @@
+import { CarnivoreNames } from '../enums/carnivoreNames';
+import { PlantNames } from '../enums/plantNames';
+
+export interface Herbivore {
+  runFromCarnivore(carnivoreName: CarnivoreNames): string;
+  eatPlant(plantName: PlantNames): string;
+}
